Banks/requirements.txt
Flask==2.0.1
flask_cors
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
pdf2image==1.16.3
psycopg2-binary==2.9.5
pypdf==3.17.1
pytesseract==0.3.10
sqlalchemy==1.4.46
Werkzeug==2.0.1
